+++ src/Credfeto.Nuget.Proxy.Logic/Services/JsonIndexTransformerBase.cs
@@ -10,6 +10,7 @@ using Credfeto.Nuget.Proxy.Index.Transformer.Interfaces;
 using Credfeto.Nuget.Proxy.Logic.Services.LoggingExtensions;
 using Credfeto.Nuget.Proxy.Models.Config;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 
 namespace Credfeto.Nuget.Proxy.Logic.Services;
 
@@ -19,14 +20,13 @@ public abstract class JsonIndexTransformerBase
     private readonly bool _indexReplacement;
     private readonly ILogger _logger;
 
-    protected JsonIndexTransformerBase(
-        ProxyServerConfig config,
-        IJsonDownloader jsonDownloader,
-        bool indexReplacement,
-        ILogger logger
+    protected JsonIndexTransformerBase(IOptions<ProxyServerConfig> config,
+                                       IJsonDownloader jsonDownloader,
+                                       bool indexReplacement,
+                                       ILogger logger
     )
     {
-        this.Config = config;
+        this.Config = config.Value;
         this._jsonDownloader = jsonDownloader;
         this._indexReplacement = indexReplacement;
         this._logger = logger;
@@ -102,7 +102,7 @@ public abstract class JsonIndexTransformerBase
 
     protected Uri GetRequestUri(string path)
     {
-        return new(this.Config.UpstreamUrls[0].CleanUri() + path);
+        return new(new Uri(this.Config.UpstreamUrls[0]).CleanUri() + path);
     }
 
     protected string ReplaceUrls(string json)
@@ -110,10 +110,10 @@ public abstract class JsonIndexTransformerBase
         return this.Config.UpstreamUrls.Aggregate(seed: json, func: this.ReplaceOneUrl);
     }
 
-    private string ReplaceOneUrl(string current, Uri uri)
+    private string ReplaceOneUrl(string current, string uri)
     {
-        string from = uri.CleanUri();
-        string to = this.Config.PublicUrl.CleanUri();
+        string from = new Uri(uri).CleanUri();
+        string to = new Uri(this.Config.PublicUrl).CleanUri();
 
         string result = current.Replace(from, to, comparisonType: StringComparison.Ordinal);