+++ src/Credfeto.Nuget.Proxy.Models/Config/ProxyServerConfig.cs
@@ -1,20 +1,37 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
+using System.Diagnostics.CodeAnalysis;
 
 namespace Credfeto.Nuget.Proxy.Models.Config;
 
 [DebuggerDisplay("Upstream: {UpstreamUrls.Count} Public: {PublicUrl} Package: {Packages}")]
-public sealed record ProxyServerConfig(
-    IReadOnlyList<Uri> UpstreamUrls,
-    Uri PublicUrl,
-    string Packages,
-    int JsonMaxAgeSeconds
-)
+public sealed class ProxyServerConfig
 {
-    public bool IsNugetPublicServer =>
-        this.UpstreamUrls.Any(h =>
-            h.DnsSafeHost.EndsWith(value: ".nuget.org", comparisonType: StringComparison.OrdinalIgnoreCase)
-        );
+    public ProxyServerConfig()
+    {
+        this.UpstreamUrls = [];
+        this.PublicUrl = "https://example.com";
+        this.Metadata = "/data/.json";
+        this.Packages = "/data/packages";
+        this.JsonMaxAgeSeconds = 60;
+    }
+
+    public List<string> UpstreamUrls { get; set; }
+
+    [SuppressMessage(category: "Microsoft.Design", checkId: "CA1056: Should ne a uri", Justification = "Not for config")]
+    public string PublicUrl { get; set; }
+
+    public string Metadata { get; set; }
+
+    public string Packages { get; set; }
+
+    public int JsonMaxAgeSeconds { get; set; }
+
+    public bool IsNugetPublicServer => this.UpstreamUrls.Exists(IsNuget);
+
+    private static bool IsNuget(string h)
+    {
+        return Uri.TryCreate(uriString: h, uriKind: UriKind.Absolute, out Uri? uri) && uri.DnsSafeHost.EndsWith(value: ".nuget.org", comparisonType: StringComparison.OrdinalIgnoreCase);
+    }
 }