+++ src/Credfeto.Nuget.Proxy.Package.Storage.FileSystem.Tests/FileSystemPackageStorageTests.cs
@@ -1,10 +1,10 @@
-using System;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 using Credfeto.Nuget.Proxy.Models.Config;
 using Credfeto.Nuget.Proxy.Package.Storage.Interfaces;
 using FunFair.Test.Common;
+using Microsoft.Extensions.Options;
 using Xunit;
 
 namespace Credfeto.Nuget.Proxy.Package.Storage.FileSystem.Tests;
@@ -16,19 +16,9 @@ public sealed class FileSystemPackageStorageTests : LoggingFolderCleanupTestBase
     public FileSystemPackageStorageTests(ITestOutputHelper output)
         : base(output)
     {
-        Uri upstream = new("https://upstream.example.org");
-        Uri publicUri = new("https://nuget.example.org");
-        ProxyServerConfig config = new(
-            [upstream],
-            PublicUrl: publicUri,
-            Packages: this.TempFolder,
-            JsonMaxAgeSeconds: 60
-        );
+        ProxyServerConfig config = new() { UpstreamUrls = ["https://upstream.example.org"], PublicUrl = "https://nuget.example.org", Packages = this.TempFolder, JsonMaxAgeSeconds = 60 };
 
-        this._packageStorage = new FileSystemPackageStorage(
-            config: config,
-            this.GetTypedLogger<FileSystemPackageStorage>()
-        );
+        this._packageStorage = new FileSystemPackageStorage(Options.Create(config), this.GetTypedLogger<FileSystemPackageStorage>());
     }
 
     [Fact]
@@ -36,10 +26,7 @@ public sealed class FileSystemPackageStorageTests : LoggingFolderCleanupTestBase
     {
         CancellationToken cancellationToken = this.CancellationToken();
 
-        byte[]? result = await this._packageStorage.ReadFileAsync(
-            sourcePath: "doesnotexist",
-            cancellationToken: cancellationToken
-        );
+        byte[]? result = await this._packageStorage.ReadFileAsync(sourcePath: "doesnotexist", cancellationToken: cancellationToken);
 
         Assert.Null(result);
     }
@@ -49,16 +36,9 @@ public sealed class FileSystemPackageStorageTests : LoggingFolderCleanupTestBase
     {
         CancellationToken cancellationToken = this.CancellationToken();
 
-        await File.WriteAllTextAsync(
-            Path.Combine(path1: this.TempFolder, path2: "file.txt"),
-            contents: "test",
-            cancellationToken: cancellationToken
-        );
+        await File.WriteAllTextAsync(Path.Combine(path1: this.TempFolder, path2: "file.txt"), contents: "test", cancellationToken: cancellationToken);
 
-        byte[]? result = await this._packageStorage.ReadFileAsync(
-            sourcePath: "file.txt",
-            cancellationToken: cancellationToken
-        );
+        byte[]? result = await this._packageStorage.ReadFileAsync(sourcePath: "file.txt", cancellationToken: cancellationToken);
 
         Assert.NotNull(result);
         Assert.Equal(expected: 4, actual: result.Length);