+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - Dependencies - Updated Microsoft.Extensions to 9.0.7
 - Dependencies - Updated SonarAnalyzer.CSharp to 10.13.0.120203
 - SDK - Updated DotNet SDK to 9.0.302
+- Separate folder for json
+- using IOptions for configuration
 ### Removed
 ### Deployment Changes